remove outside click hook from global state
