+++ components/autofill-searchbox.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState, useRef, useCallback } from 'react';
-// import { useStore } from '@/stores/store';
 import { Input } from './ui/input';
 import { Search } from 'lucide-react';
-import { useOutsideClick } from '@/stores/store';
+import { useDebounceCallback } from 'usehooks-ts';
+
 type Props = {
   searchFunction(searchText: string): void;
   setSearchInput(searchText: string): void;
@@ -21,6 +21,7 @@ export default function SingleSearchbox(props: Props) {
   const [selectedAutocompleteIndex, setSelectedAutocompleteIndex] =
     useState(-1);
   const searchRef = useRef<HTMLInputElement>(null);
+  const autocompleteRef = useRef<HTMLDivElement>(null);
 
   const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     const value = event.target.value;
@@ -126,12 +127,24 @@ export default function SingleSearchbox(props: Props) {
     }
   };
 
-  const ref = useOutsideClick(() => {
-    setShowAutocomplete(false);
-  });
+  const handleClickOutside = (event: MouseEvent) => {
+    if (
+      autocompleteRef.current &&
+      !autocompleteRef.current.contains(event.target as Node)
+    ) {
+      setShowAutocomplete(false);
+    }
+  };
+
+  useEffect(() => {
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, []);
 
   return (
-    <div className="w-full" ref={ref}>
+    <div className="w-full">
       <div className="relative">
         <form onSubmit={handleFormSubmit}>
           <Input
@@ -150,7 +163,10 @@ export default function SingleSearchbox(props: Props) {
             <Search size={15} />
           </button>
           {showAutocomplete && (
-            <div className="absolute z-50 mt-1 max-h-60 w-full overflow-y-auto rounded-md border border-zinc-500  bg-zinc-950 py-1 shadow-md">
+            <div
+              className="absolute z-50 mt-1 max-h-60 w-full overflow-y-auto rounded-md border border-zinc-500  bg-zinc-950 py-1 shadow-md"
+              ref={autocompleteRef}
+            >
               {autocompleteResults &&
                 autocompleteResults.map((result, index) => (
                   <div
