+++ stores/store.ts
@@ -1,26 +1,6 @@
 import { create } from 'zustand';
 import axiosInstance from '@/utils/axiosWrapper';
 import { toast } from 'sonner';
-import { useEffect, useRef } from 'react';
-
-export const useOutsideClick = (callback: () => void) => {
-  const sortRadioRef = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        sortRadioRef.current &&
-        !sortRadioRef.current.contains(event.target as Node)
-      ) {
-        callback();
-      }
-    };
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [callback]);
-  return sortRadioRef;
-};
 
 export interface PromoInformation {
   promoCode: string;
